% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weats.R
\docType{package}
\name{cbn}
\alias{cbn}
\alias{cbn-package}
\title{CBN}
\source{
A. Caliskan, J. J. Bryson, and A. Narayanan (2017) 'Semantics derived
        automatically from language corpora contain human-like biases'
        \emph{Science}. 356:6334 \url{http://doi.org/10.1126/science.aal4230}.
}
\description{
This package contains tools and experimental items necessary to
replicate Caliskan, Bryson, and Narayanan (2017). 'Semantics derived
        automatically from language corpora contain human-like biases'
}
