% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weats.R
\name{cbn_cosine}
\alias{cbn_cosine}
\title{Calculates Cosine Similarity Between Matrix Rows}
\source{
The original code is from the \code{cosine} function by
        Fridolin Wild (f.wild@open.ac.uk) in the \code{lsa} package.
}
\usage{
cbn_cosine(x, y = NULL)
}
\arguments{
\item{x}{A vector or a matrix (e.g., a document-term matrix).}

\item{y}{A vector with compatible dimensions to x. If NULL, use all columns of \code{x}.}
}
\value{
An \code{ncol(x)} by \code{ncol(x)} matrix of cosine similarities, a scalar
        cosine similarity, or a vector of cosine simialrities of length \code{nrow(y)}.
}
\description{
This function calculates the cosine similarity matrix between all
rows of a matrix \code{x}. When \code{x} and \code{y} are vectors
it calculates the cosine similarity between them.  When \code{x}
is a vector and \code{y} is a matrix it calculates the cosine
between \code{x} and each row of \code{y}.
}
\details{
This code is taken directly from the \code{lsa} package but adjusted to
operate rowwise.
}
