% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weats.R
\name{cbn_extract_word_vectors}
\alias{cbn_extract_word_vectors}
\title{Extract Word Vectors From Current Vector File}
\usage{
cbn_extract_word_vectors(words, verbose = FALSE, report_every = 1e+05)
}
\arguments{
\item{words}{words to get vectors for}

\item{verbose}{whether to report on progress}

\item{report_every}{how often to check in to see if we should stop}
}
\value{
a matrix with word vectors as rows
}
\description{
This function provides a more convenient wrapper for \code{extract_words}.
It uses the current vector file, whose location can be found using
\code{\link{cbn_get_vectorfile_location}} and assigned with
\code{\link{cbn_set_vectorfile_location}}.
}
