% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weats.R
\docType{data}
\name{cbn_gender_name_stats}
\alias{cbn_gender_name_stats}
\title{Gender Proportions for Names in the US Population}
\format{An object of class \code{data.frame} with 210 rows and 6 columns.}
\usage{
cbn_gender_name_stats
}
\description{
This data set contains for each name used in the study items, its
the gender proportions in the US population. It was generated
by the \code{gender} package, which uses US Social Security
Administration data.
}
\details{
The columns of the data set are \code{name}, the name,
\code{proportion_male} and \code{proportion_female}, gender (a best guess
from the proportions), and the years within which the SSA search was
performed.  This data set can merged with several of the study item sets,
but is most useful for replicating the second WEFAT study, as shown in the
replication vignette.
}
\keyword{datasets}
