% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weats.R
\docType{data}
\name{cbn_gender_name_stats_census1990}
\alias{cbn_gender_name_stats_census1990}
\title{Gender Proportions for Names in the US Population from 1990 Census}
\format{An object of class \code{data.frame} with 50 rows and 5 columns.}
\usage{
cbn_gender_name_stats_census1990
}
\description{
This data set contains gender information \emph{only} for names used in
WEFAT 2.  It is a slightly normalized version of the original Dataverse
materials.
}
\details{
The columns are \code{name}, \code{gender.score} a numerical score derived
(somehow) using -1 to be female, 0 to mean unisex, and 1 to mean male,
and \code{percentage.in.population}, \code{percentage.in.male.population},
and \code{percentage.in.female.population}.  Apparently these three are
some measure of the prevalence of the name in the US population and two
gender subpopulations.

The original materials are a tab separated file located at
\code{system.file("extdata", "censusNames1990.tsv", package = "cbn")}.

Presumably some Bayes theorem with the addition of the population gender
balance recreates the quantity of substantive interest: P(gender | name).
This has not been done.
}
\keyword{datasets}
