% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weats.R
\name{cbn_get_item_vectors}
\alias{cbn_get_item_vectors}
\title{Get Vectors for Items in a Study}
\usage{
cbn_get_item_vectors(type = c("all", "WEAT", "WEFAT"), number = 1)
}
\arguments{
\item{type}{"all" (the default), "WEAT", or "WEFAT"}

\item{number}{study number (default: 1) Ignored if \code{type} = "all"}
}
\value{
a matrix with word vectors as rows
}
\description{
Returns a matrix containing word vectors for the items used in one
of the studies (WEAT1 through WEAT10 or WEFAT1 or WEFAT2).
If \code{type} == "all" then vectors for all items used in any of the studies
is returned. Words are row names.
}
