% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weats.R
\docType{data}
\name{cbn_item_cosines}
\alias{cbn_item_cosines}
\title{Cosine Similarity for Every Pair of Study Items}
\format{An object of class \code{matrix} with 457 rows and 457 columns.}
\usage{
cbn_item_cosines
}
\description{
A matrix of cosine similarities between each item and every other one.
Uses \code{cbn_items}.
}
\keyword{datasets}
