% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weats.R
\docType{data}
\name{cbn_item_vectors}
\alias{cbn_item_vectors}
\title{Vectors for All Items Used in All Studies}
\format{An object of class \code{matrix} with 457 rows and 300 columns.}
\source{
J. Pennington, R. Socher, and C. D. Manning (2014)
        'GloVe: Global vectors for word representation'
        \url{https://nlp.stanford.edu/projects/glove/}.
}
\usage{
cbn_item_vectors
}
\description{
A 457 x 300 matrix of (row) vectors for all study items,
extracted from the 840B word Common Crawl data on Jun 30th, 2018.
}
\keyword{datasets}
