% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weats.R
\docType{data}
\name{cbn_items}
\alias{cbn_items}
\title{All Items Used in All Studies}
\format{An object of class \code{data.frame} with 730 rows and 3 columns.}
\source{
A. Caliskan, J. J. Bryson, and A. Narayanan (2017) 'Semantics derived
        automatically from language corpora contain human-like biases'
        \emph{Science}. 356:6334 \url{http://doi.org/10.1126/science.aal4230}.
}
\usage{
cbn_items
}
\description{
This data frame contains all the items used in all the studies.
It is the data source for \code{cbn_get_items}.
Most of the time you should probably use that.
}
\keyword{datasets}
