% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weats.R
\name{cbn_make_items}
\alias{cbn_make_items}
\title{Make items}
\usage{
cbn_make_items(studyname, words, conditions, roles = NULL)
}
\arguments{
\item{studyname}{Name of your study}

\item{words}{a vector of words}

\item{conditions}{a vector of condition labels (must be the same length as
\code{words})}

\item{roles}{An optional vector of role description labels (must be the same length as
\code{words}). Values are either \code{target} or \code{attribute}}
}
\value{
a set of items
}
\description{
Make items
}
