% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weats.R
\name{get_study_items}
\alias{get_study_items}
\title{Get the Items from a Study}
\usage{
get_study_items(study_type = c("WEAT", "WEFAT"), number = 1)
}
\arguments{
\item{study_type}{WEAT (default) or WEFAT}

\item{number}{study number (default: 1)}
}
\value{
a data frame of items in columns
}
\description{
Get the Items from a Study
}
