% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weats.R
\docType{data}
\name{kneer}
\alias{kneer}
\title{kneer}
\format{An object of class \code{cbn_study} (inherits from \code{data.frame}) with 215 rows and 4 columns.}
\usage{
kneer
}
\description{
Items for the Kneer Study (see also \code{\link{kneer_vecs}}).
}
\keyword{datasets}
