% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weats.R
\docType{data}
\name{kneer_vecs}
\alias{kneer_vecs}
\title{kneer_vecs}
\format{An object of class \code{matrix} with 215 rows and 300 columns.}
\usage{
kneer_vecs
}
\description{
Vectors corresponding to items for the Kneer Study
(see also \code{\link{kneer}}).
}
\keyword{datasets}
