% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weats.R
\docType{data}
\name{LIWC2001}
\alias{LIWC2001}
\title{LIWC2001}
\format{An object of class \code{list} of length 68.}
\usage{
LIWC2001
}
\description{
This is the Linguistic Inquiry and Word Count dictionary from 2001.
It has word lists for the following somewhat self-explanatory
topics:
}
\details{
Pronoun, I, We, Self, You, Other, Negate, Assent, Article, Preps, Number,
Affect, Posemo, Posfeel, Optim, Negemo, Anx, Anger, Sad, Cogmech, Cause,
Insight, Discrep, Inhib, Tentat, Certain, Senses, See, Hear, Feel, Social,
Comm, Othref, Friends, Family, Humans, Time, Past, Present, Future, Space,
Up, Down, Incl, Excl, Motion, Occup, School, Job, Achieve, Leisure, Home,
Sports, TV, Music, Money, Metaph, Relig, Death, Physcal, Body, Sexual,
Eating, Sleep, Groom, Swear, Nonfl, Fillers

Note 1: Some of the words lists are wildcarded.  The current extraction code
does not work with these, so choose some representative completions.

Note 2: This dictionary is proprietary (though how that works with federal
money is anybody's guess) so remove before publishing the package.
}
\keyword{datasets}
