% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weats.R
\name{summary.cbn_study}
\alias{summary.cbn_study}
\title{Summary Method for Study Items}
\usage{
\method{summary}{cbn_study}(object, ...)
}
\arguments{
\item{object}{A set of study items}

\item{...}{Ignored}
}
\value{
Condition names, roles (target or attribute) and N for study items
}
\description{
A summary method for study items extracted via \code{\link{cbn_get_items}}.
}
\examples{
its <- cbn_get_items("WEAT", 6)
summary(its)

}
