% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weats.R
\name{vectors_location}
\alias{vectors_location}
\title{Get the Location of the Vectors File}
\usage{
vectors_location()
}
\value{
a full path to the vectors file
}
\description{
Returns the full path to the file of word vectors.  If there is no
environment variable \code{CBN_VECTORS_LOC}, prompts to set one.
}
\details{
If you want prefer the location of your downloaded vectors to persist
across sessions, add \code{CBN_VECTORS_LOC=/Users/me/Documents/myvectors.txt}
or similar to your \code{~/.Renviron} file (creating the file if necessary).
}
