% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{weat_boot}
\alias{weat_boot}
\title{WEAT via simple item bootstrap}
\usage{
weat_boot(items, vectors, x_name, y_name, a_name, b_name, b = 300,
  se.calc = c("sd", "quantile"))
}
\arguments{
\item{items}{information about the items, typically from
\code{\link{cbn_get_items}}}

\item{vectors}{a matrix of word vectors for all the study items}

\item{x_name}{the \emph{name} of the target item condition, e.g. "Flowers"
in WEAT 1}

\item{y_name}{the \emph{name} of the target item condition, e.g. "Insects"
in WEAT 1}

\item{a_name}{the name of the first condition, e.g. "Pleasant" in
WEAT 1}

\item{b_name}{the name of the second condition, e.g. "Unpleasant" in
WEAT 1}

\item{b}{number of bootstrap samples. Defaults to 300.}

\item{se.calc}{how to compute lower and upper bounds on an approximate 95%
interval for the difference of differences of cosines statistic.
"se" (default) or "quantile".}
}
\value{
a data frame with first column the
        difference of differences of cosines statistic, the second and third
        columns the lower and upper bounds of an approximate 95% confidence
        interval from the bootstrapped statistic.  If \code{se.calc} is "quantile",
        the fourth column is the median value of the statistic across
        bootstrap samples.
}
\description{
A simple bootstrap for the WEAT calculations.  The statistic
of interest is an average difference of average differences.
}
\details{
Schematically, the statistic is the average value of

(cosine(x names, a words) - cosine(x names, b words)) -
(cosine(y names, a words) - cosine(y names, b words))

If a denotes a set of 'Pleasant' and b denotes a set of 'Unpleasant' words,
x are names of 'Insects', and y are names of 'Flowers' (as in WEAT 1)
then the statistic will take positive values when flowers are more pleasant
than insects. That is, when the degree to which flower names are more similar
to pleasant versus unpleasant words is stronger than the degree to which
insect names are more similar to pleasant versus unpleasant words.

Uncertainty is quantified by bootstrapping each set of
item vectors. That is, in each of the \code{b} bootstrap samples,
vectors in each condition (\code{a_name}, \code{b_name},
\code{x_name} and \code{y_name}) are
separately resampled with replacement, and the statistic is
computed.  The bootstrap sampling distribution of this statistic
is summarized in the outpu by an approximate
95% confidence interval defined as either twice the standard deviation of the
statistic across bootstrap samples if \code{se.calc} is "sd", or as the
0.025 and 0.975 quantiles of the bootstrap sampling distribution
if \code{se.calc} is "quantile".

If \code{se.calc} is "quantile" the data frame returned has an extra column
containing the median of the statistic in the bootstrap samples. This should not
be too far from the original statistic.

The sign of the statistic is arbitrary. If you wish to reverse
the ordering just swap the values of \code{a_name} for \code{b_name}
or \code{x_name} and \code{y_name} when calling it.

Note that this is not the statistic reported in the original paper.  This
bootstraps within each target categories (x and y) and within each attribute
category (a and b).
}
\examples{
its <- cbn_get_items("WEAT", 1)
its_vecs <- cbn_get_item_vectors("WEAT", 1)
res <- weat_boot(its, its_vecs,
                 x_name = "Flowers", y_name = "Insects",
                 a_name = "Pleasant", b_name= "Unpleasant",
                 se.calc = "quantile")
res
}
