% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{wefat_boot}
\alias{wefat_boot}
\title{WEFAT via simple item bootstrap}
\usage{
wefat_boot(items, vectors, x_name, a_name, b_name, b = 300,
  se.calc = c("sd", "quantile"))
}
\arguments{
\item{items}{information about the items, typically from
\code{\link{cbn_get_items}}}

\item{vectors}{a matrix of word vectors for the study}

\item{x_name}{the \emph{name} of the target item condition, e.g. "Careers"
in WEFAT 1}

\item{a_name}{the name of the first condition, e.g. "MaleAttributes" in
WEFAT 1 and 2}

\item{b_name}{the name of the second condition, e.g. "FemaleAttributes" in
WEFAT 1 and 2}

\item{b}{number of bootstrap samples. Defaults to 300.}

\item{se.calc}{how to compute lower and upper bounds on an approximate 95%
interval for the difference of cosines statistic. "se"
(default) or "quantile".}
}
\value{
a data frame with first column \code{x_name}, second column the
        difference of cosines statistic, third and fourth columns the
        lower and upper bounds of an approximate 95% confidence interval
        from the bootstrapped statistic. If \code{se.calc} is "quantile",
        the fifth column is the median value of the statistic across
        bootstrap samples. The data frame is sorted by the second column.
}
\description{
A simple bootstrap for the WEFAT calculations.  The statistic
of interest is the difference between the cosine of each word in condition
\code{x_name} e.g. "Careers", to the mean vector of condition \code{a_name},
e.g. "MaleAttributes" and the mean vector from condition \code{b_name},
e.g. "FemaleAttributes".
}
\details{
Uncertainty is quantified by bootstrapping each set of
item vectors. That is, in each of the \code{b} bootstrap samples,
vectors in the \code{a_name} condition and
vectors in the \code{b_name} condition are
resampled (independently) with replacement, and the difference between
the cosine of a target word and the mean of the \code{a_name}
vectors and cosine of a target word and the mean of the \code{b_name}
is recorded.  The bootstrap sampling distribution of this difference of
cosines statistic is summarized in the outpu by an approximate
95% confidence interval defined as either twice the standard deviation of the
statistic across bootstrap samples if \code{se.calc} is "sd", or as the
0.025 and 0.975 quantiles of the bootstrap sampling distribution
if \code{se.calc} is "quantile".

If \code{se.calc} is "quantile" the data frame returned has an extra column
containing the median of the statistic in the bootstrap samples. This should not
be too far from the original statistic.

The output of this function is sorted by the value of the difference of
cosines statistic. This direction is arbitrary, but if you wish to reverse
the ordering just swap the values of \code{a_name} for \code{b_name} when
calling it.

Note that this is not the statistic reported in the original paper.
}
\examples{
its <- cbn_get_items("WEFAT", 1)
its_vecs <- cbn_get_item_vectors("WEFAT", 1)
res <- wefat_boot(its, its_vecs, x_name = "Careers",
                  a_name = "MaleAttributes", b_name = "FemaleAttributes",
                  se.calc = "quantile")
}
