% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weats.R
\name{cbn_set_vectorfile_location}
\alias{cbn_set_vectorfile_location}
\title{Set the Location of the Vectors File}
\usage{
cbn_set_vectorfile_location(f)
}
\arguments{
\item{f}{path where you unzipped your vectors file}
}
\value{
Nothing
}
\description{
If you want prefer this setting to persist
across sessions, add \code{CBN_VECTORS_LOC=/Users/me/Documents/myvectors.txt}
or similar to your \code{~/.Renviron} file (creating the file if necessary).
}
